Split output into its own module
